add directories for each planned daemon
Planned daemons:
* REST Daemon
* Redis Daemon
* Worker Daemon
* Recorder Daemon

Each of the above is just an idea and could change dramatically.
Ideally I would have a number of daemons running to build a flow
of information through a system that could be expanded to be quite
complex. I would like to experiment with a microservice architecture.
Each of these daemon's has only one job.